rtmouse_i2c.cを作成してI2C関連の関数を移す
